requirements.txt
flask
flask-cors
gunicorn

pdf2docx
python-docx
docx2pdf

pypdf
PyPDF2

pdf2image
Pillow
img2pdf
reportlab

openpyxl
pandas

python-pptx

weasyprint
pdfkit
mammoth
pymupdf

requests
werkzeug